services/agent/requirements.txt
livekit-agents>=0.13.0
livekit-plugins-deepgram
livekit-plugins-elevenlabs
livekit-plugins-openai
httpx>=0.27.0
qdrant-client>=1.17.0
elasticsearch>=9.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
twilio>=9.3.0
pytest>=8.3.0
